Problem: serializable transactions aren't always enough
They can't establish more complicated data dependency scenarios,
particularly when no "dangerous structures" are formed, leading to what
can be construed as business constraint violation.

Solution: introduce serializable transaction linearization

This will fail writes if there's an outstanding SIReadLock. This may
result in false positive, so retries are highly necessary.
