+++ extensions/omni_txn/CHANGELOG.md
@@ -7,6 +7,16 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.4.0] - TBD
 
+### Added
+
+* Experimental serializable transaction linearization
+  functionality [#666](https://github.com/omnigres/omnigres/pull/666)
+
+### Fixed
+
+* `retry` was not setting up predicate locks to be tracked [#666](https://github.com/omnigres/omnigres/pull/666)
+## [0.4.0] - TBD
+
 ### Fixed
 
 * `retry` was not setting up predicate locks to be tracked [#668](https://github.com/omnigres/omnigres/pull/668)
