+++ extensions/omni_txn/omni_txn.h
@@ -1,0 +1,9 @@
+#ifndef omni_txn_h
+#define omni_txn_h
+
+#include <omni/omni_v0.h>
+
+void linearization_init(const omni_handle *handle);
+void linearize_transaction();
+
+#endif // omni_txn_h
